Move interchange, cleanup CLI, updated README
