+++ lib/interchange.js
@@ -11,11 +11,11 @@ var http    = require('http');
 var path    = require("path");
 var tmp     = require('tmp');
 
-var creators = require('./lib/firmwares.json').creators;
-var firmwares = require('./lib/firmwares.json').firmwares;
-var interchange_client = require('./lib/interchange_client');
+var creators = require('./firmwares.json').creators;
+var firmwares = require('./firmwares.json').firmwares;
+var interchange_client = require('./interchange_client');
 var ic_client = new interchange_client.Client();
-var version = require('./package.json').version;
+var version = require('../package.json').version;
 
 var fw = null; // used to hold details of the firmware we want in it.
 
@@ -33,8 +33,7 @@ Interchange.prototype.clean_temp_dir = function (tmpdir) {
 };
 
 Interchange.prototype.list_devices = function () {
-    // this function lists out all of the devices available to have firmware
-    // installed.
+    // this function lists out all of the firmwares available to be installed.
 
     console.info("\nFirmwares available for backpacks. (f) denotes a firmata version is available\n");
     _.sortBy(firmwares, "name").forEach(function(firmware) {
@@ -71,6 +70,11 @@ Interchange.prototype.get_firmware_info = function (port) {
     // attempts to connect to an interchange firmware and get the
     // installed details.
 
+    if (!port) {
+        console.error("Please provide a device path".red);
+        return;
+    }
+
     ic_client.port = port;
 
     ic_client.on("error", function(err) {
@@ -105,6 +109,37 @@ Interchange.prototype.get_firmware_info = function (port) {
             this.close();
         }.bind(this));
     })
+};
+
+Interchange.prototype.install_firmware = function (firmware, opts) {
+
+    if (!firmware) {
+        console.error("Please supply a firmware to install".red);
+        return;
+    }
+
+    var settings = {
+        board: opts.board || process.env.INTERCHANGE_BOARD || "nano",
+        port: opts.port || process.env.INTERCHANGE_PORT,
+        firmata: opts.firmata,
+        i2c_address: opts.address,
+    };
+
+    try {
+        this.check_firmware(firmware, settings, function(hex_path, tmp_dir, options) {
+
+            this.flash_firmware(hex_path, options, function() {
+                // once complete destory the tmp_dir.
+                if (tmp_dir) {
+                    this.clean_temp_dir(tmp_dir);
+                }
+            }.bind(this));
+        }.bind(this));
+    } catch (e) {
+        console.error(e);
+        return;
+    }
+
 };
 
 Interchange.prototype.set_firmware_details = function (port, opts, cb) {
@@ -232,9 +267,12 @@ Interchange.prototype.download_from_npm = function (firmware, options, cb) {
 Interchange.prototype.download_from_github = function (firmware, options, cb) {
     // downloads the firmware from the GH repo
 
-    console.info("Retrieving manifest data from GitHub".magenta);
+    var self = this;
     var manifest_uri = null;
     var base_uri = null;
+
+    console.info("Retrieving manifest data from GitHub".magenta);
+
     if (firmware.repo.indexOf('git+https') == 0) {
         base_uri = "https://raw.githubusercontent.com" + firmware.repo.substring(22) + "/master";
         manifest_uri = base_uri + "/manifest.json?" + (new Date().getTime());
@@ -252,7 +290,7 @@ Interchange.prototype.download_from_github = function (firmware, options, cb) {
 
             if (err) {
                 console.error(err);
-                clean_temp_dir(tmp_dir);
+                self.clean_temp_dir(tmp_dir);
                 throw err;
             }
 
@@ -260,7 +298,7 @@ Interchange.prototype.download_from_github = function (firmware, options, cb) {
                 var manifest = JSON.parse(fs.readFileSync(manifest_files[0].path));
             } catch (e) {
                 console.error("Manifest file incorrect");
-                clean_temp_dir(tmp_dir);
+                self.clean_temp_dir(tmp_dir);
                 throw "Manifest file error"
             }
 
@@ -294,7 +332,7 @@ Interchange.prototype.download_from_github = function (firmware, options, cb) {
                 .run(function(err, hex_files) {
                     if (err) {
                         console.error(err);
-                        clean_temp_dir(tmp_dir);
+                        self.clean_temp_dir(tmp_dir);
                         throw err;
                     }
                     cb(hex_files[0].path, tmp_dir, options);
